Fix up ADC1 handling.
Switch pin to input after reading to avoid biasing next reading.
Double ADC clock

+++ hw/controls.h
@@ -11,12 +11,11 @@ uint32_t keypad_read(void);
 char get_key(void);
 void Power_As_Input(void);
 void Normal_Power(void);
-int VOL_Read(void);
-int VOL_Taper(int vol);
-int BATT_Read(void);
-int BATT2_Read(void);
-int Temp_Read(void);
-
+uint16_t VOL_Read(void);
+uint8_t VOL_Taper(uint16_t vol);
+uint16_t BATT_Read(void);
+uint16_t BATT2_Read(void);
+uint16_t Temp_Read(void);
 
 #define KEY_PTT		'T'
 #define KEY_TOP		'~'
